+++ Runtime/Utils/ResponsiveType.cs
@@ -1,0 +1,9 @@
+namespace SaintsField.Utils
+{
+    public enum ResponsiveType
+    {
+        None,
+        Pixel,
+        Percent,
+    }
+}